Reorganized Home components
