+++ auction/src/components/Home/Jumbotron/CategoryList/CategoryList.js
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { Link } from 'react-router-dom';
 import axios from 'axios';
 
-import config from '../../config';
+import config from '../../../../config';
 
 const getCategories = async setCategories => {
     try {
