+++ auction/src/components/Home/Jumbotron/CategoryList/index.js
@@ -1,0 +1,1 @@
+export { default } from './CategoryList';
