+++ auction/src/components/Home/Jumbotron/index.js
@@ -1,0 +1,1 @@
+export { default } from './Jumbotron';
